Split catchParticleFilter.cpp into two separated files to try to reduce the test time. Increase default ctest timeout of 1500 to 2000 seconds, see: https://cmake.org/cmake/help/latest/manual/ctest.1.html#cmdoption-ctest-timeout
